+++ driver/examples/baroclinic_init.py
@@ -1,7 +1,6 @@
 from argparse import ArgumentParser
 
 import yaml
-from mpi4py import MPI
 
 from pace.driver.run import Driver, DriverConfig
 
@@ -22,12 +21,6 @@ def parse_args():
 args = parse_args()
 with open(args.config_file, "r") as f:
     driver_config = DriverConfig.from_dict(yaml.safe_load(f))
-driver = Driver(
-    config=driver_config,
-    comm=MPI.COMM_WORLD,
-)
+driver = Driver(config=driver_config)
 driver.diagnostics.store(time=driver.config.start_time, state=driver.state)
-driver.diagnostics.store_grid(
-    grid_data=driver.state.grid_data,
-    metadata=driver.state.dycore_state.ps.metadata,
-)
+driver.diagnostics.store_grid(grid_data=driver.state.grid_data)