+++ fv3core/pace/fv3core/initialization/dycore_state.py
@@ -327,7 +327,7 @@ class DycoreState:
                     extent=sizer.get_extent(dims),
                     gt4py_backend=backend,
                 )
-        state = cls(**dict_state)
+        state = cls(**dict_state)  # type: ignore
         return state
 
     @classmethod