+++ fv3core/tests/savepoint/translate/translate_map1_ppm_2d.py
@@ -36,7 +36,7 @@ class TranslateMap1_PPM_2d(TranslateDycoreFortranData2Py):
         stencil_factory: pace.dsl.StencilFactory,
     ):
         super().__init__(grid, namelist, stencil_factory)
-        self.compute_func = MapSingleFactory(stencil_factory)
+        self.compute_func = MapSingleFactory(stencil_factory)  # type: ignore
         self.in_vars["data_vars"] = {
             "q1": {"serialname": "var_in"},
             "pe1": {"istart": 3, "iend": grid.ie - 2, "axis": 1},